# morning-brief-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Morning Brief",
    "description": "Reads Reinier's Gmail and Google Calendar. Use getMorningBrief to get emails and calendar events. Use syncNow before getMorningBrief if you want the very latest data. Use createCalendarEvent to add events. Use createEmailDraft to prepare a draft reply (never sends automatically).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://morning-brief-connector.reinier-olivier.workers.dev"
    }
  ],
  "paths": {
    "/brief": {
      "get": {
        "operationId": "getMorningBrief",
        "summary": "Get emails and calendar",
        "description": "Returns categorised emails (important, new, marketing) and calendar events (today + next 7 days) from the database. Call syncNow first if you need the very latest data.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Morning brief data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "generated_at": { "type": "string" },
                    "status": { "type": "string" },
                    "emails": {
                      "type": "object",
                      "properties": {
                        "unread_count": { "type": "number" },
                        "important_messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "subject": { "type": "string" },
                              "sender_name": { "type": "string" },
                              "sender_email": { "type": "string" },
                              "received_at": { "type": "string" },
                              "body_preview": { "type": "string" },
                              "is_read": { "type": "boolean" },
                              "web_link": { "type": "string" }
                            }
                          }
                        },
                        "new_messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "subject": { "type": "string" },
                              "sender_name": { "type": "string" },
                              "sender_email": { "type": "string" },
                              "received_at": { "type": "string" },
                              "body_preview": { "type": "string" },
                              "web_link": { "type": "string" }
                            }
                          }
                        },
                        "marketing_messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "subject": { "type": "string" },
                              "sender_name": { "type": "string" },
                              "sender_email": { "type": "string" },
                              "received_at": { "type": "string" }
                            }
                          }
                        }
                      }
                    },
                    "calendar": {
                      "type": "object",
                      "properties": {
                        "today_events": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "subject": { "type": "string" },
                              "start_at": { "type": "string" },
                              "end_at": { "type": "string" },
                              "location": { "type": "string" },
                              "organiser": { "type": "string" },
                              "response_status": { "type": "string" }
                            }
                          }
                        },
                        "upcoming_events": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "subject": { "type": "string" },
                              "start_at": { "type": "string" },
                              "end_at": { "type": "string" },
                              "location": { "type": "string" },
                              "organiser": { "type": "string" },
                              "response_status": { "type": "string" }
                            }
                          }
                        }
                      }
                    },
                    "warnings": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "operationId": "syncNow",
        "summary": "Sync latest emails and calendar from Gmail",
        "description": "Pulls the latest emails and calendar events from Gmail and Google Calendar into the database. Call this before getMorningBrief when the user wants up-to-the-minute data.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sync result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "overall": { "type": "string" },
                    "email": {
                      "type": "object",
                      "properties": {
                        "status": { "type": "string" },
                        "itemsProcessed": { "type": "number" }
                      }
                    },
                    "calendar": {
                      "type": "object",
                      "properties": {
                        "status": { "type": "string" },
                        "itemsProcessed": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events": {
      "post": {
        "operationId": "createCalendarEvent",
        "summary": "Create a Google Calendar event",
        "description": "Creates a new event in Reinier's Google Calendar. Always confirm the details with the user before calling this.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subject", "startTime", "endTime", "timezone"],
                "properties": {
                  "subject": { "type": "string", "description": "Event title" },
                  "startTime": { "type": "string", "description": "ISO 8601 e.g. 2026-10-10T09:00:00" },
                  "endTime":   { "type": "string", "description": "ISO 8601 e.g. 2026-10-10T10:00:00" },
                  "timezone":  { "type": "string", "description": "IANA timezone e.g. Africa/Johannesburg" },
                  "location":  { "type": "string" },
                  "body":      { "type": "string", "description": "Event description" },
                  "attendees": { "type": "array", "items": { "type": "string" }, "description": "Email addresses of attendees" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Event created successfully" }
        }
      }
    },
    "/drafts": {
      "post": {
        "operationId": "createEmailDraft",
        "summary": "Save a Gmail draft (does NOT send)",
        "description": "Creates a draft email in Reinier's Gmail Drafts folder. This NEVER sends the email — the user must open Gmail and send it manually. Always confirm recipient and content before calling.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subject", "body", "toRecipients"],
                "properties": {
                  "subject":      { "type": "string" },
                  "body":         { "type": "string", "description": "HTML or plain text body" },
                  "toRecipients": { "type": "array", "items": { "type": "string" }, "description": "Recipient email addresses" },
                  "ccRecipients": { "type": "array", "items": { "type": "string" } },
                  "importance":   { "type": "string", "enum": ["low", "normal", "high"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Draft saved to Gmail Drafts folder" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token — enter in ChatGPT Authentication field only, never in this schema"
      }
    }
  },
  "security": [{ "BearerAuth": [] }]
}